refactor(frontend): extract shared usePackSeparatelyUpdater hook
OrderFormBody.tsx and PackSeparatelySelector.tsx each carried an identical
ref-to-latest-sections + lookup-then-dispatch block, differing only in
variable names — a fix to the merged/remainder update logic had to be
applied twice by hand. Moved it into packSeparately.ts as one hook both
components call.
